Reduce instance struct size 184 -> 176B
Nice 8 byte saving, probably could save more space still, I don't really
like the big list of emission and bsdfs in every instance struct, but I
haven't figured out a better way yet. Instances can have their own
unique set of materials, distinct from the actual object by the means of
material overrides in the scene description.
I suppose those materials could all be stuffed into the original object,
and then I'd just have a pointer to the object and a material offset,
but that would complicate the already ugly scene loading code even more.

+++ src/utils/loaders/sceneloader.c
@@ -752,7 +752,7 @@ void apply_materials_to_instance(struct renderer *r, struct instance *instance, 
 	} else {
 		// Single graph, map it to every material in a mesh.
 		const struct bsdfNode *node = parseBsdfNode(r->prefs.assetPath, r->state.file_cache, &r->scene->storage, overrides);
-		for (size_t i = 0; i < instance->bsdf_count; ++i) {
+		for (int i = 0; i < instance->bsdf_count; ++i) {
 			instance->bsdfs[i] = node;
 		}
 	}
